+++ packages/web-twig/tests/DependencyInjection/ConfigurationTest.php
@@ -17,7 +17,10 @@ spirit_web_twig:
     paths:                []
     paths_alias:          spirit
     spirit_css_class_prefix: null
-    html_syntax_lexer:    true\n
+    html_syntax_lexer:    true
+    icons:
+        paths:                []
+        alias:                icons-assets\n
 CONFIG;
 
         $this->assertEquals($reference, $dumper->dump(new Configuration()));